backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Jira
jira==3.6.0

# AI/ML
transformers==4.36.2
torch==2.1.2
numpy==1.26.3

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing (optional, for development)
pytest==7.4.4
httpx==0.26.0

# Week 2: Authentication & Billing
pyjwt[crypto]==2.8.0
stripe==7.4.0
email-validator==2.1.0
